requirements.txt
flask
flask-cors
numpy
opencv-python
pillow